Do not specify midas credential in CMakeLists.txt
Midas API key should be considered as a password and should not be
encoded in clear.

Additonally, it overwrites the factory credentials when the extension is built and package
every night.

Instead, user could set the environment variables MIDAS_PACKAGE_EMAIL and
MIDAS_PACKAGE_API_KEY. See [1] for more details.

Specifying the url is also not needed, the expected value is
already set in SlicerConfig.cmake and will be available after
calling "find_package(Slicer)"

[1] http://wiki.slicer.org/slicerWiki/index.php/Documentation/Nightly/Developers/FAQ#How_to_cache_API_credentials_.3F