Simplify inclusion of VTK package
Since (1) VTK_DIR will be set by either the user or by SlicerConfig.cmake
(after calling"find_package(Slicer)") and (2) calling find_package for the
same package multiple time should not be an issue, this commit removes the
code doing conditional inclusion of VTK package.

Additionally, in case the package can NOT be found, considering the
"REQUIRED" argument is provided,  the find_package will complain. For that
reason, the code associated with "if(NOT VTK_DIR)" will never be executed
and is removed by this commit.